requirements.txt
streamlit
flask
sqlalchemy
passlib
python-dotenv
requests
pandas
numpy
matplotlib
scipy
agno
